% Generated by roxygen2 (4.0.1): do not edit by hand
\name{an}
\alias{an}
\title{Expected total time on genealogy for W-F coalescent}
\usage{
an(n)
}
\arguments{
\item{n}{The sample size.}
}
\value{
The expected total time on the tree (in units of 4N generations)
}
\description{
Expected total time on genealogy for W-F coalescent
}

