% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bistruct}
\alias{bistruct}
\title{The "structure" algorithm for biallelic data and K = 2}
\usage{
bistruct(data, burn = 2000, m = 2000)
}
\arguments{
\item{data}{A data frame/matrix.  The first column must be sampling locations, and the remaining columns are genotypes}

\item{burn}{Number of MCMC iterations for burn in.}

\item{m}{Number of MCMC generations after the burn in step.}
}
\value{
An array of assignment probabilities.  These are the probability that each individual belongs to population 1.
}
\description{
The "structure" algorithm for biallelic data and K = 2
}
\details{
See help(genoData) for more detail on the format that data must take.
}
\note{
This is the most basic implementation of the structure algorithm possible.  The implementation here
was in response to homework assignment number 4 in Carlos Bustamante's Biometry 694 course at Cornell.
}
\examples{
data(genoData)
genoData.assigned = bistruct(genoData)
}
\references{
Pritchard, J. K., M. Stephens, and P. Donnelly (2000) Inference of Population Structure Using Multilocus Genotype Data. Genetics 155:945-959
}

