% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{coalTree}
\alias{coalTree}
\title{coalTree}
\description{
A bifurcating genealogy representing the history of a sample of size n
}
\details{
a coalTree is a list with the following elements:
nodes: An array of 2n-1 integers.  nodes[i] is the ancestor of the i-th node.
	 The root is labeled with a -1
times: An array of 2n-1 times, corresponding to the values in nodes.
tmrca: The tmrca of the sample, equivalent to times[2n-1]
ttot:  The total time on the tree
}

