% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deltaq}
\alias{deltaq}
\title{General formula for delta q for 2-allele model}
\usage{
deltaq(q, wAA, wAa, waa)
}
\arguments{
\item{q}{Current allele frequency}

\item{wAA}{Fitness of AA genotype}

\item{wAa}{Fitness of Aa genotype}

\item{waa}{Fitness of aa genotype}
}
\value{
The change in allele frequency in an infinitely-large Wright-Fisher population
}
\description{
General formula for delta q for 2-allele model
}

