% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fixtrajdominance}
\alias{fixtrajdominance}
\title{Returns a trajectory of a mutation that has fixed due to selection}
\usage{
fixtrajdominance(N, s, h)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}

\item{h}{The domimance of the mutant allele}
}
\value{
A list with the allele frequency trakectory plus the number of attempts needed before a mutation fixed
}
\description{
Returns a trajectory of a mutation that has fixed due to selection
}

