% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{genoData}
\alias{genoData}
\title{genoData}
\format{\preformatted{'data.frame':	20 obs. of  11 variables:
 $ V1 : int  1 1 2 2 1 1 2 2 2 2 ...
 $ V2 : int  1 0 0 1 0 0 1 0 0 0 ...
 $ V3 : int  0 0 0 0 0 1 1 1 0 1 ...
 $ V4 : int  0 0 1 1 1 1 1 0 1 1 ...
 $ V5 : int  1 0 0 0 1 1 1 1 1 1 ...
 $ V6 : int  0 0 0 0 1 1 1 1 1 1 ...
 $ V7 : int  0 1 0 0 1 0 1 1 1 0 ...
 $ V8 : int  0 0 0 1 0 0 0 0 0 0 ...
 $ V9 : int  0 0 1 0 0 0 1 0 0 0 ...
 $ V10: int  1 0 1 0 1 1 1 0 0 1 ...
 $ V11: int  0 0 1 0 1 1 0 0 1 1 ...
}}
\usage{
genoData
}
\description{
A data frame consisting of hypothetical genotypes.
The first column represents sampling locations (location 1 or location 2).
The remaining columns are markers, and rows are individual haplotypes.  Data values are 0 or 1,
and represent the two alleles. The data are assumed to be diploid, e.g., pairs of haplotypes
on adjacent rows are the same individual.
This file was originally provided by
Carlos Bustamante for his Biometry 694 course at Cornell University
at some point in the mid-2000s.
}
\keyword{datasets}

