% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HOC.VGVE.dist}
\alias{HOC.VGVE.dist}
\title{Distribution of VG and VE under Gaussian HOC model}
\usage{
HOC.VGVE.dist(N, mu, sigmaMU, sigmaE, sigmaS, ngens, nreps)
}
\arguments{
\item{N}{The number of diploids}

\item{mu}{The mutation rate (per gamete, per generation)}

\item{sigmamu}{The standard deviation in the effect size of new mutations}

\item{sigmae}{The standard deviation in random effects added to phenotype}

\item{sigmas}{The standard deviation in the Gaussuian fitness function}

\item{ngens}{The number of generations to simulate}

\item{nreps}{The number of replicates to simulate}
}
\value{
A data frame with the VG and VE components of phenotype from each replicate
}
\description{
Distribution of VG and VE under Gaussian HOC model
}
\details{
uses replicate to call HOCsim repeatedly
}

