% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HOCsim}
\alias{HOCsim}
\title{Simulate a quantitative trait under a House-of-Cards mutation model}
\usage{
HOCsim(N, mu, sigmamu, sigmae, sigmas, ngens)
}
\arguments{
\item{N}{The number of diploids}

\item{mu}{The mutation rate (per gamete, per generation)}

\item{sigmamu}{The standard deviation in the effect size of new mutations}

\item{sigmae}{The standard deviation in random effects added to phenotype}

\item{sigmas}{The standard deviation in the Gaussuian fitness function}

\item{ngens}{The number of generations to simulate}
}
\value{
A matrix.  Columns 1 and 2 are the alleles for each diploid.  Column 3 is the random effect.  Phenotype is therefore given by rowSums(result)
}
\description{
Simulate a quantitative trait under a House-of-Cards mutation model
}
\details{
Tersely: the mutation model is Kingman's house-of-cards with Gaussian effects.  The inspiration for this function is Turelli.
}
\examples{
\dontrun{
N=5e2
mu=0.001
sige=0.15
sigmu=0.5
sigs=1
pop = HOCsim(N,mu,sigmu,sige,sigs,8*N)
}
}
\references{
Kingman, J. (1978). A simple model for the balance between selection and mutation. Journal of Applied Probability, 15, 1-12.

Turelli, M. (1984). Heritable genetic variation via mutation-selection balance: Lerch's zeta meets the abdominal bristle. Theoretical Population Biology, 25(2), 138-193.
}

