% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isdescendant}
\alias{isdescendant}
\title{Asks if an individual on a tree is a descendant of a branch}
\usage{
isdescendant(tree, ind, branch)
}
\arguments{
\item{tree}{A simulated coalescent history}

\item{ind}{The integer label of a sample}

\item{branch}{The index of the branch}
}
\value{
TRUE or FALSE
}
\description{
Asks if an individual on a tree is a descendant of a branch
}

