% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mutate}
\alias{mutate}
\title{Infinitely-many sites mutation}
\usage{
mutate(tree, n, theta)
}
\arguments{
\item{tree}{A simulated coalescent history}

\item{n}{The sample size in tree}

\item{theta}{4Nu}
}
\description{
Infinitely-many sites mutation
}

