% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mutateS}
\alias{mutateS}
\title{Add specific number of mutations to a tree}
\usage{
mutateS(tree, n, S)
}
\arguments{
\item{tree}{A simulated coalescent history}

\item{n}{The sample size in tree}

\item{S}{The number of mutations to add}
}
\description{
Add specific number of mutations to a tree
}

