% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neutral.trajectory}
\alias{neutral.trajectory}
\title{Trajectory of a new neutral mutation in a diploid Wright-Fisher population of size N}
\usage{
neutral.trajectory(N)
}
\arguments{
\item{N}{The diploid population size}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new neutral mutation in a diploid Wright-Fisher population of size N
}
\examples{
#Fate of a new mutant in a population of N = 100 diploids
n.traj = neutral.trajectory(100)
}

