% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Nhaps}
\alias{Nhaps}
\title{calculate the number of haplotypes (unique sequences) in the sample}
\usage{
Nhaps(d)
}
\arguments{
\item{d}{A simulated data set}
}
\description{
calculate the number of haplotypes (unique sequences) in the sample
}

