% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pfix}
\alias{pfix}
\title{Estimate fixation prob and time to fixation by simulation}
\usage{
pfix(q0, N, wAA, wAa, waa, nsims)
}
\arguments{
\item{q0}{initial mutant allele frequency}

\item{N}{Population size}

\item{wAA}{Fitness of AA genotype}

\item{wAa}{Fitness of Aa genotype}

\item{waa}{Fitness of aa genotype}

\item{nsims}{The number of replicates to run}
}
\value{
A list.  prob = estimate of fixation probability.  times = distribution of fixation times
}
\description{
Estimate fixation prob and time to fixation by simulation
}

