% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pfixdominance}
\alias{pfixdominance}
\title{Estimate fixation probability of a new mutation under models with arbitrary dominance}
\usage{
pfixdominance(N, s, h, howmany)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}

\item{h}{The domimance of the mutant allele}

\item{howmany}{The number of simulation to run}
}
\value{
An estimate of the fixation probability
}
\description{
Estimate fixation probability of a new mutation under models with arbitrary dominance
}

