% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pfixgenic}
\alias{pfixgenic}
\title{Estimate fixation probability under genic/additive selection}
\usage{
pfixgenic(N, s, howmany)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}

\item{howmany}{The number of simulation to run}
}
\value{
An estimate of the fixation probability
}
\description{
Estimate fixation probability under genic/additive selection
}

