% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pickbranch}
\alias{pickbranch}
\title{Pick a random branch from a tree}
\usage{
pickbranch(tree, n, r)
}
\arguments{
\item{tree}{A simualted coalescent history}

\item{n}{The sample size corresponding to tree}

\item{r}{A random number corresponding to runif(1,0,tree$ttot)}
}
\value{
The index of the randomly-chosen branch
}
\description{
Pick a random branch from a tree
}
\note{
This is an internal function
}

