% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Rcode4teaching}
\alias{Rcode4teaching}
\alias{Rcode4teaching-package}
\title{Rcode4teaching.}
\description{
Stuff I use to generate examples, etc., for lectures
}
\author{
Kevin R. Thornton \email{krthornt@uci.edu}
}
\references{
\url{https://github.com/molpopgen/Rcode4teaching}
}

