% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtraj}
\alias{rtraj}
\title{Simulate random trajectory under selection starting from q0.}
\usage{
rtraj(q0, N, wAA, wAa, waa)
}
\arguments{
\item{q0}{initial mutant allele frequency}

\item{N}{Population size}

\item{wAA}{Fitness of AA genotype}

\item{wAa}{Fitness of Aa genotype}

\item{waa}{Fitness of aa genotype}
}
\value{
A random allele frequency trajectory
}
\description{
Simulate random trajectory under selection starting from q0.
}

