% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simWFtree}
\alias{simWFtree}
\title{Simulate a genealogy under the equilibrium WF model for a sample of size n.}
\usage{
simWFtree(n)
}
\arguments{
\item{n}{The sample size}
}
\value{
A list containing:
nodes: An array of 2n-1 integers.  nodes[i] is the ancestor of the i-th node.
	 The root is labeled with a -1
times: An array of 2n-1 times, corresponding to the values in nodes.
tmrca: The tmrca of the sample, equivalent to times[2n-1]
ttot:  The total time on the tree
}
\description{
Simulate a genealogy under the equilibrium WF model for a sample of size n.
}

