% Generated by roxygen2 (4.0.1): do not edit by hand
\name{textbook.dominance}
\alias{textbook.dominance}
\title{Trajectory of a new  mutation subject to selection with arbitrary dominance in a diploid Wright-Fisher population of size N}
\usage{
textbook.dominance(N, s, h)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}

\item{h}{The dominance of the mutant allele}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new  mutation subject to selection with arbitrary dominance in a diploid Wright-Fisher population of size N
}
\details{
Fitnesses are 1, 1+hs, and 1+2s for genotypes AA, Aa, and aa, respectively.  Thus, a is the mutant allele, and h = 1 is the genic/additive/codominant case.
}
\examples{
d.traj = textbook.dominance(100,0.01,1)
}

