% Generated by roxygen2 (4.0.1): do not edit by hand
\name{textbook.genic}
\alias{textbook.genic}
\title{Trajectory of a new  mutation subject to genic/additive selection in a diploid Wright-Fisher population of size N}
\usage{
textbook.genic(N, s)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new  mutation subject to genic/additive selection in a diploid Wright-Fisher population of size N
}
\details{
Fitnesses are 1, 1+s, and 1+2s for genotypes AA, Aa, and aa, respectively.  Thus, a is the mutant allele.
}
\examples{
g.traj = textbook.genic(100,0.01)
}

