% Generated by roxygen2 (4.0.1): do not edit by hand
\name{textbook.overdominance}
\alias{textbook.overdominance}
\title{Trajectory of a new  mutation subject to overdomimant selection in a diploid Wright-Fisher population of size N}
\usage{
textbook.overdominance(N, s1, s2, ifreq, ngen)
}
\arguments{
\item{N}{The diploid population size}

\item{s1}{The selection coefficient applied to AA}

\item{s2}{The selection coefficient applied to aa}

\item{ifreq}{The initial frequency of the a allele.}

\item{ngen}{The number of generations to iterator the function.}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new  mutation subject to overdomimant selection in a diploid Wright-Fisher population of size N
}
\details{
Fitnesses are 1-s1, 1, and 1-s2 for genotypes AA, Aa, and aa, respectively.
}
\examples{
od.traj = textbook.overdominance(100,0.1,0.1,0.2,100)
}

