% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WFdominance}
\alias{WFdominance}
\title{Trajectory of a new  mutation subject to selection with arbitrary dominance in a diploid Wright-Fisher population of size N}
\usage{
WFdominance(N, s, h)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}

\item{h}{The dominance of the mutant allele}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new  mutation subject to selection with arbitrary dominance in a diploid Wright-Fisher population of size N
}

