% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WFgenic}
\alias{WFgenic}
\title{Trajectory of a new  mutation subject to genic/additive selection in a diploid Wright-Fisher population of size N}
\usage{
WFgenic(N, s)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new  mutation subject to genic/additive selection in a diploid Wright-Fisher population of size N
}

