% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WFoverdominance}
\alias{WFoverdominance}
\title{Trajectory of a new  mutation subject to overdomimant selection in a diploid Wright-Fisher population of size N}
\usage{
WFoverdominance(N, s1, s2, ngen, ifreq)
}
\arguments{
\item{N}{The diploid population size}

\item{s1}{The selection coefficient applied to AA}

\item{s2}{The selection coefficient applied to aa}

\item{h}{The dominance of the mutant allele}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new  mutation subject to overdomimant selection in a diploid Wright-Fisher population of size N
}

