% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WFsample}
\alias{WFsample}
\title{Simulate a sample of size n under the idealized Wright-Fisher model}
\usage{
WFsample(n, theta)
}
\arguments{
\item{n}{The sample size}

\item{theta}{The mutation rate (4Nu)}
}
\value{
An object of class coalSample
}
\description{
Simulate a sample of size n under the idealized Wright-Fisher model
}

