% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WFsampleS}
\alias{WFsampleS}
\title{Simulate a sample of size n under the idealized Wright-Fisher model}
\usage{
WFsampleS(n, S)
}
\arguments{
\item{n}{The sample size}

\item{S}{The number of mutations to place in the sample}
}
\value{
An object of class coalSample
}
\description{
Simulate a sample of size n under the idealized Wright-Fisher model
}

