% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WFtrajectory}
\alias{WFtrajectory}
\title{Trajectory of a new neutral mutation in a diploid Wright-Fisher population of size N}
\usage{
WFtrajectory(N)
}
\arguments{
\item{N}{The diploid population size}
}
\value{
The allele frequency in each generation until fixation or loss
}
\description{
Trajectory of a new neutral mutation in a diploid Wright-Fisher population of size N
}

