% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pfixgenic}
\alias{pfixgenic}
\title{Estimate fixation probability of a new mutation under genic/additive selection}
\usage{
pfixgenic(N, s, howmany)
}
\arguments{
\item{N}{The diploid population size}

\item{s}{The selection coefficient}

\item{howmany}{The number of simulation to run}
}
\value{
An estimate of the fixation probability
}
\description{
Estimate fixation probability of a new mutation under genic/additive selection
}
\details{
Fitnesses are 1, 1+s, and 1+2s for genotypes AA, Aa, and aa, respectively.  Thus, a is the mutant allele.
}
\examples{
pfix.genic = pfixgenic(100,0.1,10)
}

