% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rtraj}
\alias{rtraj}
\title{Simulate random trajectory under selection starting from q0.}
\usage{
rtraj(q0, N, wAA, wAa, waa, ngens = -1)
}
\arguments{
\item{q0}{initial mutant allele frequency}

\item{N}{Population size}

\item{wAA}{Fitness of AA genotype}

\item{wAa}{Fitness of Aa genotype}

\item{waa}{Fitness of aa genotype}

\item{ngens}{The number of generations to iterate. If -1, the function runs until fixation or loss occurs.  If fixation/loss occurs prior to ngens, the trajectory up until that generation is returned}
}
\value{
A random allele frequency trajectory
}
\description{
Simulate random trajectory under selection starting from q0.
}

