% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simWFtree}
\alias{simWFtree}
\title{Simulate a genealogy under the equilibrium WF model for a sample of size n.}
\usage{
simWFtree(n)
}
\arguments{
\item{n}{The sample size}
}
\value{
A class of type coalTree
}
\description{
Simulate a genealogy under the equilibrium WF model for a sample of size n.
}

